Finalized picture code for edit and display
Added in more code to support proper displaying of all user pictures.
WHEW.  What a pain jquery can be.  Made it so that user can properly
edit the titles of the pictures taken as well as view them full screen
using a popup menu that appears everytime they click a picture
thumbnail.