Copied over Panel funcitonality to iOS and integrated a "fix"
Copied over all codes and image assets for the side pane into the iOS
side project.  Also forced a change page for when the user clicks the
checkbox in settings as iOS apparently has a bug that makes the entire
window the checkbox immediately after checking it making it impossible
to change pages with the checkbox being enabled.  Weird.